refactor(all): Renamed 'files' to same nomenclature and reorganized 'server.ts'
