+++ server/src/services/pizza.service.ts
@@ -3,7 +3,7 @@
 import {
     DuplicatedProductError,
     InvalidArgValuesError,
-} from "../errors/errors";
+} from "../errors/errors.error";
 import { pizzaModel } from "../data/mongodb/models/Pizza";
 // import { PaginationParams } from "../types/types";
 interface IPizza {
